requirements.txt
flask>=2.3.0
pyyaml>=6.0
openapi-spec-validator>=0.7.1
requests>=2.31.0
jinja2>=3.1.2
click>=8.1.7
python-frontmatter>=1.1.0
openai>=1.12.0
anthropic>=0.20.0
python-dotenv>=1.0.0
rich>=13.7.0
